modify how models are defined and loaded
Previously models were magically loaded in using fs, simplified
that so know instead of importing models, they are defined directly.

The models/index.js still exists so that all models can be loaded in at
once and associations can be defined in one location.

This change also allows for named importing of models as well as
importing models individually